feat: 재사용을 위한 LabelInput 컴포넌트 분리
